requirements.txt
python-dotenv
opencv-python
azure-cognitiveservices-vision-face
azure-cognitiveservices-vision-computervision
